requirements.txt
flask>=3.0.0
flask-cors>=3.0.0
ddgs>=4.0.0
python-dotenv>=1.0.0
google-genai>=0.1.0
groq>=0.9.0
ollama>=0.3.0
requests>=2.31.0
edge-tts>=6.1.0
flask-sqlalchemy>=3.1.0
psycopg2-binary>=2.9.0
pg8000>=1.30.0
gunicorn>=20.1.0